+++ frontend/src/app/result/page.tsx
@@ -1,7 +1,7 @@
 import type { SearchParams } from '~/types/search-params'
 import { redirect } from 'next/navigation'
-import MapClient from '~/components/features/map/MapClient'
-import RestaurantList from '~/components/features/restaurant/RestaurantList'
+import MapClient from '~/components/features/map/map-client'
+import RestaurantList from '~/components/features/restaurant/restaurant-list'
 import { fetchRestaurants } from '~/services/fetch-restaurants'
 import { getApiKey } from '~/services/get-api-key'
 import { parseAndValidateCoordinates } from '~/services/parse-and-validate-coords'
