+++ frontend/src/components/features/account/auth/Auth.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import { useMediaQuery } from '~/hooks/useMediaQuery'
-import { AuthDialog } from './AuthDialog'
-import { AuthDrawer } from './AuthDrawer'
+import { AuthDialog } from './auth-dialog'
+import { AuthDrawer } from './auth-drawer'
 
 export function Auth() {
   const isDesktop = useMediaQuery('(min-width: 768px)')
