+++ frontend/src/components/features/account/delete/delete-account-responsive.tsx
@@ -1,14 +1,14 @@
 'use client'
 
 import { useMediaQuery } from '~/hooks/useMediaQuery'
-import DeleteAccountDialog from './Dialog'
-import DeleteAccountDrawer from './Drawer'
+import DeleteAccountDialog from './delete-account-dialog'
+import DeleteAccountDrawer from './delete-account-drawer'
 
 export default function DeleteAccount() {
   const isDesktop = useMediaQuery('(min-width: 768px)')
   return (
     <>
-      { isDesktop ? <DeleteAccountDialog /> : <DeleteAccountDrawer /> }
+      {isDesktop ? <DeleteAccountDialog /> : <DeleteAccountDrawer />}
     </>
   )
 }
