+++ frontend/src/components/features/map/Map.tsx
@@ -6,8 +6,8 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import { MapContainer } from 'react-leaflet'
 import { calculateCardPosition } from '~/utils/calculate-card-position'
 import { createLeafletOptions } from '~/utils/create-leaflet-options'
-import MapContent from './MapContent'
-import MapRestaurantCard from './MapRestaurantCard'
+import MapContent from './map-content'
+import MapRestaurantCard from './map-restaurant-card'
 import 'leaflet/dist/leaflet.css'
 import '@maptiler/sdk/dist/maptiler-sdk.css'
