+++ frontend/src/components/features/map/map-restaurant-card.tsx
