+++ frontend/src/components/features/map/midpoint-marker.tsx
