+++ frontend/src/components/features/map/restaurant-marker.tsx
