+++ frontend/src/components/features/restaurant/favorite-button.tsx
