+++ frontend/src/components/features/restaurant/restaurant-sidebar.tsx
@@ -6,9 +6,9 @@ import Link from 'next/link'
 import { useEffect, useRef } from 'react'
 import { Button } from '~/components/ui/button'
 import { cn } from '~/utils/cn'
-import RestaurantCard from './RestaurantCard'
-import RestaurantListHeader from './RestaurantListHeader'
-import RestaurantPagination from './RestaurantPagination'
+import RestaurantCard from './restaurant-card'
+import RestaurantListHeader from './restaurant-list-header'
+import RestaurantPagination from './restaurant-pagination'
 
 interface RestaurantSidebarProps {
   restaurants: RestaurantListItem[]
