+++ frontend/src/components/features/station/search/buttons/remove-form-button.tsx
